autorelease: Fix "v" handling for tags (#26123)
When we create the release, we need to use the same format as the
original tag. If we always prepend "v" then GH will create a new copy of
the tag under that name.

The most straightforward thing to do seemed to be to make `TAG` always
be the original tag, and strip the leading "v" in the cases where we
wanted it stripped.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3017480542