push-to-mirrors: Set Upstream-Ref in mirror commits, and use when creating branches (#36850)
One drawback to the current functioning is that a new branch that's not
based on trunk will still be mirrored as if it's based on trunk, meaning
the first commit will seem to be reverting everything added to trunk
since its actual branch point.

The fix for this has two parts:

* All new commits in the mirror repos will have an "Upstream-Ref" footer
  indicating the monorepo commit it was mirrored from. Probably we
  should have been doing that all along, but we didn't think of it.
* When creating a new branch in the mirror, use the "Upstream-Ref" on
  existing commits to try to find the actual branch point (falling back
  to trunk if no match is found).

Then we also add some configuration options in case an external user of
the action doesn't want this functionality, or wants to limit the
number of revisions scanned in a repo with a lot of history.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8663988147

Upstream-Ref: Automattic/jetpack@836f04f42f01d4d1298b7a1a6c6505e160a3d0d7