push-to-mirrors: Check for misordered Actions runs (#43913)
One problem with this action is that, if runs for source-repo commits
are misordered, then the mirror repo branch may wind up on an older
commit than the source repo.

We can avoid that by making use of the Upstream-Ref footers: if the
mirror repo branch we're about to update has an Upstream-Ref footer
referencing a commit that's a descendant of the commit we're about to
mirror, we should skip mirroring it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/15683779209

Upstream-Ref: Automattic/jetpack@7f8d4e80052aaab9f7f6d0be0eb9cc1c6b70c091